AvailabilityCheck logger method and shared spec
