+++ lib/topological_inventory/providers/common/operations/sources_api_client.rb
@@ -25,14 +25,20 @@ module TopologicalInventory
             endpoints.find(&:default)
           end
 
-          def fetch_authentication(source_id, default_endpoint = nil)
+          def fetch_authentication(source_id, default_endpoint = nil, authtype = nil)
             endpoint = default_endpoint || fetch_default_endpoint(source_id)
             return if endpoint.nil?
 
             endpoint_authentications = api.list_endpoint_authentications(endpoint.id.to_s).data || []
             return if endpoint_authentications.empty?
 
-            auth_id = endpoint_authentications.first.id
+            auth_id = if authtype.nil?
+                        endpoint_authentications.first&.id
+                      else
+                        endpoint_authentications.detect { |a| a.authtype = authtype }&.id
+                      end
+            return if auth_id.nil?
+
             fetch_authentication_with_password(auth_id)
           end
