+++ spec/spec_helper.rb
@@ -7,6 +7,8 @@ require "bundler/setup"
 require "topological_inventory/providers/common"
 require "webmock/rspec"
 
+Dir["./spec/support/**/*.rb"].each {|f| require f}
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
